Create FlyingFish.java
Write a class named FlyingFish, which has 4 fields: name (a String), weight (an int), speed (a
float), and id (an int). It should have:
 Three constructors: the default constructor assigns “unnamed”, 3, and 0.0 to name, weight, and
speed, respectively; the second constructor takes a String as the argument, and assigns the string,
3, and 0.0 to the three fields; the third constructor takes a String, an int, and a float as the
arguments, and assigns them to the three fields.
 For each FlyingFish object created, it has a unique id in the increasing order, starting from 1 for
the first object. This id assignment is also implemented in the constructors.
 An overridden toString() method, which prints out all the 4 fields (print format is shown in the
below examples).
 A main method that, for each argument to the main method, creates a FlyingFish object (using the
argument as the String input to the constructor) and then prints itself out.
Two test examples:
java FlyingFish Mike Jack
>name: Mike, weight: 3, speed: 0.0, id: 1
>name: Jack, weight: 3, speed: 0.0, id: 2
java FlyingFish test
>name: test, weight: 3, speed: 0.0, id: 1