Create Vagetable.java
Write a program named VegeSequence.java, which has the below types:
 A Vegetable interface that has one abstract method toString()
 A Tomato class that implements Vegetable and defines toString() as printing out “I am a
tomato!”
 A Lettuce class and Cabbage class implemented similarly as Tomato
 The VegeSequence class itself, whose members include an array of 3 vegetables initialized with a
Tomato, a Lettuce, and a Cabbage. It also has an inner class VegeIterator that is capable of
iterating through all the elements in the vegetable array by providing the end(), current(), and
next() methods. In the main method, create a VegeSequence object and a VegeIterator object,
and use the latter to iterate the former and print out all the vegetable names.
The program should work like below:
$ java VegeSequence
$ Tomato
$ Lettuce
$ Cabbage
Hint: This is similar to the SeqSelector example shown in the lecture slides.