Create ReversedWord.java
Write a program named ReversedWords that can print out all the words in a given sentence in the
reversed order. For example, if the below sentence is given:
Java has the feature of "write once, run anywhere" (WORA), meaning that compiled Java code can
run on all platforms without recompilation...
The output of the program would be:
recompilation without platforms all on run can code Java compiled that meaning WORA anywhere
run once write of feature the has Java
Note that:
 Only pure words are printed out (in the reverse order). All punctuations are removed.
 You can hardcode the above sentence in your program, so that when running “java
ReversedWords” on the command line the above output will appear.
 As a hint to implement this program, you can split the sentence with a certain regex, and store all
the words in a stack.