Create JavaFileLister.java
Write a program named JavaFilesLister, which lists all the .java files in the current working
directory.
The program should work like below (if it is executed in a directory that contains all the files for this
homework). It prints out all the .java files of this homework; and the order doesn’t matter:
$ java JavaFilesLister
$ JavaFilesLister.java
$ ReversedWords.java
$ VegeSequence.java
$ SortedRabbits.java
$ WordCharFrequencies.java
Hint: You will need to create a File object for the current directory, and use its list(FilenameFilter)
method. You will need to implement the FilenameFilter.